+++ scripts/make_summary.py
@@ -40,14 +40,18 @@ def assign_carriers(n: pypsa.Network) -> None:
 
 
 def assign_locations(n: pypsa.Network) -> None:
-    for c in n.iterate_components(n.one_port_components):
-        c.df["location"] = c.df.bus.map(n.buses.location)
+    for c in n.components[n.one_port_components]:
+        if c.static.empty:
+            continue
+        c.static["location"] = c.static.bus.map(n.buses.location)
 
-    for c in n.iterate_components(n.branch_components):
-        c_bus_cols = c.df.filter(regex="^bus")
+    for c in n.components[n.branch_components]:
+        if c.static.empty:
+            continue
+        c_bus_cols = c.static.filter(regex="^bus")
         locs = c_bus_cols.apply(lambda c: c.map(n.buses.location)).sort_index(axis=1)
         # Use first location that is not "EU"; take "EU" if nothing else available
-        c.df["location"] = locs.apply(
+        c.static["location"] = locs.apply(
             lambda row: next(
                 (loc for loc in row.dropna() if loc != "EU"),
                 "EU",